adding lifecycle spec
